feat: frise infobulle ready
