+++ app/App.js
@@ -15,22 +15,25 @@ import NPS from "./src/services/NPS/NPS";
 import { Sentry } from "react-native-sentry";
 import { NeedUpdateContextProvider } from "./src/context/needUpdate";
 import { InfoModalProvider } from "./src/components/InfoModal";
+import { SafeAreaProvider } from "react-native-safe-area-context";
 
 if (!__DEV__) {
   Sentry.config("https://9f0bd8f8af8444eea9f470d00a1bb411@sentry.fabrique.social.gouv.fr/54").install();
 }
 
 const App = () => (
-  <NeedUpdateContextProvider>
-    <DiaryNotesProvider>
-      <DiaryDataProvider>
-        <InfoModalProvider>
-          <Router />
-          <NPS />
-        </InfoModalProvider>
-      </DiaryDataProvider>
-    </DiaryNotesProvider>
-  </NeedUpdateContextProvider>
+  <SafeAreaProvider>
+    <NeedUpdateContextProvider>
+      <DiaryNotesProvider>
+        <DiaryDataProvider>
+          <InfoModalProvider>
+            <Router />
+            <NPS />
+          </InfoModalProvider>
+        </DiaryDataProvider>
+      </DiaryNotesProvider>
+    </NeedUpdateContextProvider>
+  </SafeAreaProvider>
 );
 
 export default App;
