+++ app/package.json
@@ -27,8 +27,10 @@
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.17.7",
     "@react-native-community/art": "react-native-community/art",
+    "@react-native-community/blur": "^4.2.0",
     "@react-native-community/checkbox": "^0.5.6",
     "@react-native-community/datetimepicker": "^3.5.0",
+    "@react-native-community/hooks": "^2.8.1",
     "@react-native-community/masked-view": "^0.1.10",
     "@react-native-community/netinfo": "^5.9.9",
     "@react-native-community/push-notification-ios": "^1.10.1",
